Add files via upload
Some comment lines were recorded in Turkish. Please take this into consideration for the translation.

Additionally, if there are variable names that you cannot define, they may be written according to Turkish characters, which may be a problem in UTF-8.

If you need to change these, take the language structure into consideration.